Cleanup and compiler compatibility fixes.
git-svn-id: https://ahruman.googlecode.com/svn/trunk@56 23014adc-a008-11dd-9a14-c1645b63b0c6

+++ Graphics/SGSceneGraph/Scene Graph/Tags/SGConditionalTag.m
@@ -227,7 +227,7 @@ static NSString *Description(id value)
 		}
 		else
 		{
-			NSLog(@"%s: invalid operation %i.", __FUNCTION__, inOperation);
+			NSLog(@"%s: invalid operation %i.", __FUNCTION__, (int)inOperation);
 		}
 	}
 }
